requirements.txt
langchain==0.0.352
langchain-community==0.0.13
openai==1.6.1
faiss-cpu==1.7.4
pinecone-client==2.2.4
pypdf==3.17.4
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0